Refactoring - spice things up, scroll to item after edit / delete and applied animation

+++ app/Http/Controllers/CarController.php
@@ -23,7 +23,8 @@ class CarController extends Controller
         $car->image = $image_path;
         $car->save();
 
-        $request->session()->flash('status', "$car->name added Successfully!"); //success toast
+        //success toast, I use the car id to scroll to the car that's just been added and apply animation on it
+        $request->session()->flash('status', ["message"=>"$car->name added Successfully!","car_id"=>$car->id]);
 
         return redirect('car/list');
 
@@ -45,9 +46,7 @@ class CarController extends Controller
                 'image' => $image_validation_rules,
                 'price' => 'required',
                 'name' => 'required'
-        ]);
-
-        
+        ]);        
 
     }
 
@@ -55,8 +54,9 @@ class CarController extends Controller
         
         $car = Car::find($car_id);
 
-        $car  =  $car ?  $car : false;
-
+        $car  =  $car ?  $car : false;     
+        $request->session()->forget('status'); //avoids 'car/daleted successfully' message being shown if the user goes back without hitting save
+        $request->session()->save();
         return view('car/edit',['car' => $car]);
 
     }
@@ -65,17 +65,18 @@ class CarController extends Controller
 
         $this->validateData($request,'edit'); 
         
-        $car = Car::find($car_id);       
+        $car = Car::find($car_id);   
+        
+        //if a malicious user tempers with the car id and submit an invalid car id, direct them to car list
+        if(!$car) return redirect('car/list');
+        
         if($request->hasFile($this->imageInputName)){ //new image
 
              ///delete old image
              Storage::disk('public_images')->delete($car->image);
            
             $new_path = $this->storeImage($request);
-            $car->image = $new_path;
-           
-            
-            
+            $car->image = $new_path;         
         }
 
         $car->price = $request->price;
@@ -83,8 +84,8 @@ class CarController extends Controller
         
         $car->save();
         
-
-        $request->session()->flash('status', "$car->name updated Successfully!"); //success toast
+         //success toast, I use the car id to scroll to the car that's just been updated and apply animation on it
+        $request->session()->flash('status', ["message"=>"$car->name updated Successfully!",'car_id'=>$car->id]); //success toast
 
         return redirect('car/list');
         
@@ -109,8 +110,11 @@ class CarController extends Controller
 
    public function delete(Request $request){
 
-        $car_id = $request->car_id;
+        $car_id = $request->car_id;        
+
         $car = Car::find($car_id);
+        //if a malicious user tempers with the car id and submit an invalid car id, redirect them to car list
+        if(!$car)  return response()->json(['url'=>url('car/list')]);
 
         $car_name = $car->name;
 
@@ -119,7 +123,7 @@ class CarController extends Controller
 
         $car->delete();
 
-        $request->session()->flash('status', "$car_name deleted Successfully!"); //success toast
+        $request->session()->flash('status', ["message"=>"$car_name deleted Successfully!"]); //success toast
 
         //returns json response to ajax
         return response()->json(['url'=>url('car/list')]);
@@ -127,7 +131,7 @@ class CarController extends Controller
    }
 
     public function list(){        
-
+       
         $cars = DB::table('cars')->orderBy('created_at', 'desc')->get();
         return view('car.list', ['cars' => $cars]);
